pyproject.toml
[tool.poetry]
name = "configs"
version = "1.0.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.2.2"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
